Make prototype of Serial_::begin consistent with implementation

+++ cores/arduino/USB/USBAPI.h
@@ -69,7 +69,7 @@ private:
 	RingBuffer *_cdc_rx_buffer;
 public:
 	void begin(uint32_t baud_count);
-	void begin(unsigned long, uint8_t);
+	void begin(uint32_t, uint8_t);
 	void end(void);
 
 	virtual int available(void);
